Finalize build-config generalization
