+++ .github/workflows/on-push.yaml
@@ -44,10 +44,8 @@ jobs:
         run: |
           TASKS=$(node ./tools/nrwl-utils/gen-matrix.js $GITHUB_REF)
           echo "::set-output name=TASKS::$TASKS"
-
-  # Prints the dependencies of the project into a public graph. To be used later in
-  # with other jobs. Generates a static artifact
-  gen-dep-graph:
+  # Generates static files for the internal app.
+  gen-static:
     name: gen dep-graph
     timeout-minutes: 5
     if: "!contains(github.event.head_commit.message, 'skip ci') && github.ref == 'refs/heads/main'"
@@ -80,6 +78,7 @@ jobs:
         run: nx report
       - name: nx:workspace-lint
         run: nx workspace-lint
+      # gen dep-graph
       - name: gen dep-graph
         run: nx dep-graph --file=dist/dep-graph/index.html
       - name: upload dep-graph
@@ -88,85 +87,25 @@ jobs:
           name: dep-graph
           path: dist/dep-graph
           retention-days: 3
-
-  # Runs compodoc and uploads the generated documentation to the artifacts. To be used later
-  # to build out an "internal" app.
-  gen-docs:
-    name: gen docs
-    timeout-minutes: 5
-    if: "!contains(github.event.head_commit.message, 'skip ci') && github.ref == 'refs/heads/main'"
-    runs-on: ubuntu-latest
-    steps:
-      - name: checkout repo
-        uses: actions/checkout@v2
-        with:
-          # get lfs items
-          lfs: true
-          # fetch all commits, slower but required to pass to nx
-          fetch-depth: 0
-      - name: setup node
-        uses: actions/setup-node@v2
-        with:
-          node-version: '16.10.0'
-          cache: 'npm'
-      - name: cache dependencies
-        id: cache
-        uses: actions/cache@v2
-        with:
-          path: ./node_modules
-          key: modules-${{ hashFiles('package-lock.json') }}
-      - name: install dependencies
-        if: steps.cache.outputs.cache-hit != 'true'
-        run: npm ci --ignore-scripts
-      - name: install global nx
-        run: npm install -g @nrwl/cli
+      # gen docs
       - name: gen-docs
-        run: nx run docs-generator:html
+        run: npx compodoc -p tools/docs-generator/tsconfig.docs.json --output=dist/docs/
       - name: upload dep-graph
         uses: actions/upload-artifact@v2
         with:
           name: docs
           path: dist/docs
           retention-days: 3
-  # Runs tooling to build the "config" script. Can be used for multiple builds within
-  # the pipeline.
-  build-config:
-    name: build config
-    timeout-minutes: 5
-    if: "!contains(github.event.head_commit.message, 'skip ci') && github.ref == 'refs/heads/main'"
-    runs-on: ubuntu-latest
-    steps:
-      - name: checkout repo
-        uses: actions/checkout@v2
-        with:
-          # get lfs items
-          lfs: false
-          # fetch all commits, slower but required to pass to nx
-          fetch-depth: 0
-      - name: setup node
-        uses: actions/setup-node@v2
-        with:
-          node-version: '16.10.0'
-          cache: 'npm'
-      - name: cache dependencies
-        id: cache
-        uses: actions/cache@v2
-        with:
-          path: ./node_modules
-          key: modules-${{ hashFiles('package-lock.json') }}
-      - name: install dependencies
-        if: steps.cache.outputs.cache-hit != 'true'
-        run: npm ci --ignore-scripts
-      - name: install global nx
-        run: npm install -g @nrwl/cli
+      # gen build-config
       - name: build-config
-        run: nx run build-config:generate
+        run: nx workspace-generator generate-config --path=./dist/config.json
       - name: upload build-config
         uses: actions/upload-artifact@v2
         with:
           name: build-config
           path: dist/config.json
           retention-days: 3
+
   # Gathers both the docs generated by compodoc under the `gen-docs` job, and the
   # dep-graph assets generated by `nx` under the `gen-dep-graph` job
   # and deploys them to github-pages.
